test(trie/rocksdb): add integration tests and benchmarks (read path pending)
